+++ CHANGELOG.md
@@ -34,6 +34,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Bijective numeration".
 - Tests for "Plight of the fellowship of the ring".
 - Tests for "DDCG Mapper".
+- Tests for "Queneau numbers".
 
 ## [1.15.0] - 2022-11-30
 ### Added